+ progress with event driven design, balance refactored from long to double, moved messages codes from exceptions to props, a lot of main services work

+++ src/main/java/net/kravuar/giveaways/domain/dto/GiveawayFormDTO.java
@@ -10,7 +10,7 @@ import java.time.ZonedDateTime;
 @Data
 public class GiveawayFormDTO {
     @Min(1)
-    private final Long amount;
+    private final Double amount;
     @Min(1)
     private final Long usages;
     @NotNull
